+++ crates/ruma-events/src/lib.rs
@@ -164,6 +164,8 @@ pub mod key;
 #[cfg(feature = "unstable-msc3488")]
 pub mod location;
 pub mod marked_unread;
+#[cfg(feature = "unstable-msc4278")]
+pub mod media_preview_config;
 #[cfg(feature = "unstable-msc4171")]
 pub mod member_hints;
 #[cfg(feature = "unstable-msc1767")]