+++ crates/ruma-events/CHANGELOG.md
@@ -4,6 +4,9 @@ Improvements:
 
 - Add support for ThumbHash as an alternative to BlurHash for MSC2448 in 
   `ImageInfo` and `VideoInfo`.
+- Add implementation of [MSC 4278](https://github.com/matrix-org/matrix-spec-proposals/pull/4278)
+  using `MediaPreviewConfigEventContent` for the stable version and `MediaPreviewConfigEventContent`
+  for the unstable one, both are only implemented as `GlobalAccountData` events as of right now.
 
 # 0.30.2